Move data keyWriter to separate file
